+++ Moq.Dapper.Test/DapperQueryAsyncTest.cs
@@ -4,7 +4,6 @@ using System.Data.Common;
 using System.Linq;
 using Dapper;
 using NUnit.Framework;
-using static Moq.Dapper.Test.DapperQueryTest;
 
 namespace Moq.Dapper.Test
 {
@@ -106,5 +105,16 @@ namespace Moq.Dapper.Test
                 Assert.That(match.Count, Is.EqualTo(1));
             }
         }
+    }
+
+    public class ComplexType
+    {
+        public int IntegerProperty { get; set; }
+        public string StringProperty { get; set; }
+        public Guid GuidProperty { get; set; }
+        public DateTime DateTimeProperty { get; set; }
+        public DateTime? NullableDateTimeProperty { get; set; }
+        public int? NullableIntegerProperty { get; set; }
+        public byte[] ByteArrayPropery { get; set; }
     }
 }
