+++ Moq.Dapper.Test/DapperQueryTest.cs
@@ -113,16 +113,5 @@ namespace Moq.Dapper.Test
                 Assert.That(match.Count, Is.EqualTo(1));
             }
         }
-
-        public class ComplexType
-        {
-            public int IntegerProperty { get; set; }
-            public string StringProperty { get; set; }
-            public Guid GuidProperty { get; set; }
-            public DateTime DateTimeProperty { get; set; }
-            public DateTime? NullableDateTimeProperty { get; set; }
-            public int? NullableIntegerProperty { get; set; }
-            public byte[] ByteArrayPropery { get; set; }
-        }
     }
 }
